Measuring rolling and cumulative average, total sample length
Prior code was measuring period wrong assuming clock has a seconds
pendulum

This commit assumes that the photo-sensor is interrupted once per second,
that is, one swing (or half period) of the pendulum.

It saves a sample_size number of samples (sample being the difference
between current millis() and last millis()). Cumulative average of the
samples is also measured, sampled up to the maximum value of an unsigned
long. Rolling average can be queried.

Currently displays the measured length of the last interval of
sample_size samples, converted from milliseconds to seconds.